STASH

+++ src/datastream.jl
@@ -1,9 +1,9 @@
 # This file is a part of H3DPolaris.jl, licensed under the MIT License (MIT).
 
 struct PolarisEventHeader
-    nhits::Int  # Event multiplicity (number of interactions)
-    ts_s  # Event time, seconds
-    ts_ms  # Event time, milliseconds
+    nhits::Int32    # Event multiplicity (number of interactions)
+    ts_s::Int32     # Event time, seconds
+    ts_ms::Int32    # Event time, milliseconds
 end
 
 export PolarisEventHeader
@@ -12,7 +12,7 @@ Base.time(header::PolarisEventHeader) = header.ts_s + header.ts_ms * 1E-3
 
 
 function Base.read(src::IO, ::Type{PolarisEventHeader})
-    nhits = ntoh(read(src, UInt8)) # multiplicity, number of interactions
+    nhits = ntoh(read(src, UInt8))  # multiplicity, number of interactions
     ts_s = ntoh(read(src, UInt32))
     ts_ms = ntoh(read(src, UInt32))
 
@@ -22,11 +22,11 @@ end
 
 
 struct PolarisHit
-    detno::Int # Detector number
-    x::Int  # Position in µm,
-    y::Int  # Position in µm,
-    z::Int  # Position in µm,
-    edep::Int # Energy deposition in eV
+    detno::Int32 # Detector number
+    x::Int32  # Position in µm,
+    y::Int32  # Position in µm,
+    z::Int32  # Position in µm,
+    edep::Int32 # Energy deposition in eV
 end
 
 export PolarisHit
